+++ Changelog.md
@@ -1,1 +1,2 @@
 # Changelog
+- Add option in the mod settings to toggle incident filter